+++ tests/sdk/data_factory.py
@@ -525,7 +525,7 @@ class AssetsFactory:
             metadata=metadata,
         )
 
-    def create_compute_plan(self, key=None, tag="", clean_models=False, metadata=None):
+    def create_compute_plan(self, key=None, tag="", name="Test compute plan", clean_models=False, metadata=None):
         return substra.sdk.schemas.ComputePlanSpec(
             key=key or random_uuid(),
             traintuples=[],
@@ -533,6 +533,7 @@ class AssetsFactory:
             aggregatetuples=[],
             testtuples=[],
             tag=tag,
+            name=name,
             metadata=metadata,
             clean_models=clean_models,
         )