+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- BREAKING CHANGE: add mandatory name field to compute plan (#180)
+
 ## [0.20.0](https://github.com/owkin/substra/releases/tag/0.20.0) - 2022-05-09
 
 ### Changed